fix lacks

+++ user-management/src/main/java/com/toyota/usermanagement/service/UserService.java
@@ -6,6 +6,8 @@ import com.toyota.usermanagement.entity.User;
 import com.toyota.usermanagement.exception.NotFoundException;
 import com.toyota.usermanagement.repository.RoleRepository;
 import com.toyota.usermanagement.repository.UserRepository;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +18,7 @@ import java.util.stream.Collectors;
 
 @Service
 public class UserService{
+    private static Logger logger = LogManager.getLogger(UserService.class);
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
     private final PasswordEncoder encoder;
@@ -30,10 +33,12 @@ public class UserService{
         String username = user.getUsername();
 
         if(username.isEmpty()){
+            logger.warn("Username must not be empty.");
             return "Username mustn't be empty!";
         }
 
         if(userRepository.existsByUsername(username)){
+            logger.warn("Username is already taken.");
             return "Username is already taken! Please choose another one.";
         }
 
@@ -46,11 +51,14 @@ public class UserService{
 
         userRepository.save(user);
 
+        logger.info("User added successfully");
         return "User added successfully";
     }
 
     public String saveRole(Role role) {
         roleRepository.save(role);
+
+        logger.info("Role added successfully");
         return "Role added successfully";
     }
 
@@ -58,6 +66,7 @@ public class UserService{
         List<User> users = userRepository.findAll();
         List<UserDto> userDtos = new ArrayList<>();
 
+        logger.info("Getting all users");
         return users.stream()
                 .map(UserDto::convert)
                 .collect(Collectors.toList());
@@ -66,6 +75,7 @@ public class UserService{
     public UserDto getUserById(String userId) {
         User user = userRepository.findById(userId).orElseThrow(()-> new NotFoundException("User not found."));
 
+        logger.info("Getting user with id: " + userId);
         return UserDto.convert(user);
 
     }
@@ -77,11 +87,21 @@ public class UserService{
         user.setPassword(encoder.encode(userDto.password()));
 
         User updatedUser = userRepository.save(user);
+
+        logger.info("Updating user object");
         return UserDto.convert(updatedUser);
     }
 
-    public void deleteUser(String userId) {
-        userRepository.deleteById(userId);
+    public String deleteUser(String userId) {
+        if(userRepository.existsById(userId)){
+            userRepository.deleteById(userId);
+            logger.info("User has deleted with id: " + userId);
+            return "User has deleted with id: " + userId;
+        }
+        else{
+            logger.warn("User not found with id: " + userId);
+            return "User not found with id: " + userId;
+        }
     }
 
 }
